Combine diagnostics and formattings tests in 1 suite
